Добавлен Dockerfile и внесены изменения в docker-compose.yaml
