+++ docker-compose.yaml
@@ -1,6 +1,6 @@
 services:
 
-  sast-manger:
+  sast-manager:
     build: sast-manager
     image: manager-img
     container_name: manager-con
@@ -32,14 +32,14 @@ services:
       - sast-db-data:/var/lib/postgresql/data
   #    restart: "on-failure"
 
-  sast-broker:
-    build: sast-broker
-    image: broker-img
-    container_name: broker-con
-    ports:
-      - "7070:7070"
-    depends_on:
-      - sast-manger
+#  sast-broker:
+#    build: sast-broker
+#    image: broker-img
+#    container_name: broker-con
+#    ports:
+#      - "7070:7070"
+#    depends_on:
+#      - sast-manager
   #    restart: "on-failure"
 
   sast-agent:
@@ -49,7 +49,7 @@ services:
     ports:
       - "6060:6060"
     depends_on:
-      - sast-manger
+      - sast-manager
   #    restart: "on-failure"
 
   sast-ui:
@@ -57,7 +57,10 @@ services:
     image: ui-img
     container_name: ui-con
     ports:
-      - "3000:3000"
+      - "3000:80"
     depends_on:
       - sast-manager
 #    restart: "on-failure"
+
+volumes:
+  sast-db-data:
