+++ sast-manager/src/main/resources/application.properties
@@ -1,16 +1,17 @@
 server.port=8080
 spring.jpa.hibernate.ddl-auto=none
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.properties.hibernate.format_sql=true
 spring.sql.init.mode=always
 logging.level.org.springframework.orm.jpa=INFO
 logging.level.org.springframework.transaction=INFO
 logging.level.org.springframework.transaction.interceptor=TRACE
 logging.level.org.springframework.orm.jpa.JpaTransactionManager=DEBUG
+
 spring.datasource.driverClassName=org.postgresql.Driver
 spring.datasource.url=${SPRING_DATASOURCE_URL}
 spring.datasource.username=${SPRING_DATASOURCE_USERNAME}
 spring.datasource.password=${SPRING_DATASOURCE_PASSWORD}
+
 spring.jpa.open-in-view=false
 jwt.secret = jl1929MyfUeRgmRVHx07Qg/nicNEeRf5JJJF+s46mcA=
