+++ sast-ui/nginx.conf
@@ -1,0 +1,14 @@
+server {
+    listen 80;
+    server_name localhost;
+
+    root /usr/share/nginx/html;
+    index index.html;
+
+    location / {
+        # Проверяем, существует ли запрашиваемый файл
+        try_files $uri /index.html;
+    }
+
+    error_page 404 /index.html;
+}
