+++ sast-ui/src/components/Login.js
@@ -1,40 +1,42 @@
 import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
 import '../styles/Auth.css';
 
 const AUTH_URL = 'http://localhost:8080/auth';
 
-function Registration() {
-    const [formData, setFormData] = useState({ username: '', password: '', confirmPassword: '' });
+function Login() {
+    const [formData, setFormData] = useState({ username: '', password: '' });
     const [message, setMessage] = useState('');
+    const navigate = useNavigate();
 
     const handleChange = (e) => {
         const { name, value } = e.target;
         setFormData({ ...formData, [name]: value });
     };
 
-    const handleSignUp = async () => {
-        if (formData.password !== formData.confirmPassword) {
-            setMessage('Пароли не совпадают');
-            return;
-        }
-
+    const handleSignIn = async () => {
         try {
-            const response = await axios.post(`${AUTH_URL}/sign-up`, {
-                username: formData.username,
-                password: formData.password
-            }, {
+            const response = await axios.post(`${AUTH_URL}/login`, formData, {
                 headers: { 'Content-Type': 'application/json' },
             });
-            setMessage(`Регистрация успешна! Токен: ${response.data.token}`);
+            console.log(response.data); // Лог успешного ответа
+            setMessage(`Вход успешен! Токен: ${response.data.token}`);
+
+            // Сохранение токена в localStorage
+            localStorage.setItem('token', response.data.token);
+
+            // Перенаправление после успешного входа
+            navigate('/get');
         } catch (error) {
-            setMessage('Ошибка регистрации: ' + (error.response?.data?.message || error.message));
+            console.error(error.response?.data); // Лог ответа с ошибкой
+            setMessage('Ошибка входа: ' + (error.response?.data?.message || error.message));
         }
     };
 
     return (
         <div className="auth-container">
-            <h2>Регистрация</h2>
+            <h2>Вход</h2>
             <div>
                 <label>Имя пользователя:</label>
                 <input
@@ -55,20 +57,13 @@ function Registration() {
                     required
                 />
             </div>
-            <div>
-                <label>Подтверждение пароля:</label>
-                <input
-                    type="password"
-                    name="confirmPassword"
-                    value={formData.confirmPassword}
-                    onChange={handleChange}
-                    required
-                />
-            </div>
-            <button onClick={handleSignUp}>Зарегистрироваться</button>
+            <button onClick={handleSignIn}>Войти</button>
+            <button onClick={() => navigate('/auth/register')} className="secondary-button">
+                Регистрация
+            </button>
             {message && <p>{message}</p>}
         </div>
     );
 }
 
-export default Registration;
+export default Login;
